feat(rfid): wire RFID→Excel (#73) + EPC decoder tests (#74)
#73 – RFID result merged into scan record before Excel write
- RfidScanCoordinator.OnBarcodeScannedAsync returns Task<RfidValidationResult?>
- SessionViewModel awaits coordinator, merges rfidResult into VerificationRecord,
  appends rfidSuffix to status message; removed event-handler pattern
- VerificationRecord: Block K (7 RFID fields) added
- TruCheckCompatibleSchema, DataMatrix2DMapper, ISO15416Mapper: Block K columns

#74 – xUnit test project for EPC decode logic
- DeviceInterface.Tests/DeviceInterface.Tests.csproj (new)
- DeviceInterface.Tests/Rfid/EpcParserTests.cs: 13 tests covering live SGTIN-96
  captures, partition detection, SSCC-96 stub, unknown header, malformed input,
  and GTIN-14 check digit validation
- All 13 tests pass (dotnet test, net8.0, Linux)

Bugs fixed en route:
- Sgtin96Decoder: GTIN-14 was assembled as GCP+ItemRef+check (wrong); correct
  GS1 order is indicator+GCP+item_body+check (indicator is MSB decimal digit
  of ItemReference, prepended before GCP in the GTIN-14 output)
- Sgtin198Decoder: same GTIN-14 ordering bug fixed
- UnknownSchemeDecoder: implicit int→byte narrowing (CS0266) fixed
- DeviceInterface.csproj: conditional Compile Remove for SDK-dependent files
  (DataManSdkClient, DeviceSession, NetworkDiscoverer, AsReaderP35UEpcReader,
  EpcReaderFactory) so pure-C# EPC decode core builds on Linux/CI without DLLs

Test vector notes: live-B serial, live-C serial, and defect-repro GCP in the
JSON test-vector file differ from direct bit-level extraction; the JSON values
are wrong (vector was created from an earlier Python decoder with a separate
bug). Test assertions use bit-correct values.

+++ vtccp/DeviceInterface/Rfid/Schemes/UnknownSchemeDecoder.cs
@@ -24,7 +24,7 @@ public static class UnknownSchemeDecoder
     /// </summary>
     public static ParsedEpc Decode(byte[] epcBytes)
     {
-        byte header = epcBytes.Length > 0 ? epcBytes[0] : 0;
+        byte header = epcBytes.Length > 0 ? epcBytes[0] : (byte)0;
 
         Dbg($"Unrecognised EPC header 0x{header:X2} — raw: {Convert.ToHexString(epcBytes)}");
